requirements.txt
fastapi
requests
uvicorn
requests_oauthlib
pydantic
pyhumps